move generateTestCaCerts() to test-runner utils.ts file
